refactor: transferred DetailRow to a shareable component
