+++ src/app/admin/application/[id]/_components/ContactInfoCard.tsx
@@ -1,4 +1,4 @@
-import { DetailRow } from "./DetailRow";
+import { DetailRow } from "@/components/detail-row";
 
 interface ContactInfoCardProps {
   emailAddress: string;
