+++ src/app/admin/members/[id]/_components/MemberDetailsCards.tsx
@@ -1,4 +1,5 @@
 import { Building2, MapPin, User } from "lucide-react";
+import { DetailRow } from "@/components/detail-row";
 import { Badge } from "@/components/ui/badge";
 import { Card, CardContent } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
@@ -21,29 +22,6 @@ interface MemberDetailsCardsProps {
   member: MemberDetailsByBusinessMemberId;
 }
 
-function DetailRow({
-  label,
-  value,
-  valueClassName,
-}: {
-  label: string;
-  value: React.ReactNode;
-  valueClassName?: string;
-}) {
-  return (
-    <div className="flex flex-col gap-1">
-      <span className="font-medium text-muted-foreground text-xs uppercase tracking-wide">
-        {label}
-      </span>
-      <span
-        className={`font-semibold text-base leading-tight ${valueClassName ?? ""}`}
-      >
-        {value}
-      </span>
-    </div>
-  );
-}
-
 function RepresentativeField({
   label,
   value,
@@ -54,16 +32,12 @@ function RepresentativeField({
   valueClassName?: string;
 }) {
   return (
-    <div className="flex flex-col gap-1">
-      <span className="font-medium text-muted-foreground text-xs uppercase tracking-wide">
-        {label}
-      </span>
-      <span
-        className={`font-semibold text-sm leading-tight ${valueClassName ?? ""}`}
-      >
-        {value}
-      </span>
-    </div>
+    <DetailRow
+      label={label}
+      size="sm"
+      value={value}
+      valueClassName={valueClassName}
+    />
   );
 }
