+++ src/app/membership/application/_components/forms/Step4Review.tsx
@@ -2,6 +2,7 @@ import { Building2, MapPin, User } from "lucide-react";
 import Image from "next/image";
 import { useEffect, useState } from "react";
 import type { useMembershipStep4 } from "@/app/membership/application/_hooks/useMembershipStep4";
+import { DetailRow } from "@/components/detail-row";
 import { Badge } from "@/components/ui/badge";
 import { Card, CardContent } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
@@ -14,29 +15,6 @@ interface StepProps {
   form: ReturnType<typeof useMembershipStep4>["form"];
   applicationData: MembershipApplicationData;
   sectors: Sector[];
-}
-
-function DetailRow({
-  label,
-  value,
-  valueClassName,
-}: {
-  label: string;
-  value: React.ReactNode;
-  valueClassName?: string;
-}) {
-  return (
-    <div className="flex flex-col gap-1">
-      <span className="font-medium text-muted-foreground text-xs uppercase tracking-wide">
-        {label}
-      </span>
-      <span
-        className={cn("font-semibold text-base leading-tight", valueClassName)}
-      >
-        {value}
-      </span>
-    </div>
-  );
 }
 
 function formatEnumValue(value?: string): string {
@@ -82,16 +60,12 @@ function RepresentativeField({
   valueClassName?: string;
 }) {
   return (
-    <div className="flex flex-col gap-1">
-      <span className="font-medium text-muted-foreground text-xs uppercase tracking-wide">
-        {label}
-      </span>
-      <span
-        className={cn("font-semibold text-sm leading-tight", valueClassName)}
-      >
-        {value}
-      </span>
-    </div>
+    <DetailRow
+      label={label}
+      size="sm"
+      value={value}
+      valueClassName={valueClassName}
+    />
   );
 }
